Concise commit message:
```text
feat: introduce SVE2 kernel declarations for BitNet inference

- Add `extern "C"` declarations for `bitnet_dot_product_sve2` and `ggml_bitnet_transform_sve2` kernels in `bitnet.h`.
- Update `BitNetModel::generate` placeholder to indicate SVE2 readiness.
- Include `stdint.h` to support fixed-width integer types used in kernel signatures.
```

+++ app/src/main/cpp/bitnet/bitnet.h
@@ -6,7 +6,7 @@
 
 // SVE2 Kernel Declarations
 extern "C" {
-    float bitnet_dot_product_sve2(const int8_t* weights, const int8_t* activations, int n);
+    float bitnet_dot_product_sve2(const int8_t* weights, const int8_t* activations, int64_t n);
     void ggml_bitnet_transform_sve2(const void* src, void* dst, int elements);
 }
